requirements.txt
#pip3 install torch==1.9.0+cu111 torchvision==0.10.0+cu111 torchaudio==0.9.0 -f https://download.pytorch.org/whl/torch_stable.html

pytorch-lightning==1.5.10
numpy==1.19.2
tokenizers==0.11.4
omegaconf==2.1.1